PYTHON-5805 Move test_init_kms_connect_callback to test_kms_connect.py
