+++ test/test_encryption.py
@@ -19,7 +19,6 @@ from __future__ import annotations
 import asyncio
 import base64
 import copy
-import dataclasses
 import http.client
 import inspect
 import json
@@ -70,7 +69,6 @@ from pymongo.encryption_options import (
     _HAVE_PYMONGOCRYPT,
     AutoEncryptionOpts,
     HTTPProxyKMSConnect,
-    KMSConnectContext,
     RangeOpts,
     StringOpts,
     TextOpts,
@@ -237,28 +235,6 @@ class TestAutoEncryptionOpts(PyMongoTestCase):
         ctx = _kms_ssl_contexts["kmip"]
         self.assertEqual(ctx.check_hostname, True)
         self.assertEqual(ctx.verify_mode, ssl.CERT_REQUIRED)
-
-    @unittest.skipUnless(_HAVE_PYMONGOCRYPT, "pymongocrypt is not installed")
-    def test_init_kms_connect_callback(self):
-        opts = AutoEncryptionOpts({}, "k.d")
-        self.assertIsNone(opts._kms_connect_callback)
-
-        def callback(context):
-            raise AssertionError("not called")
-
-        opts = AutoEncryptionOpts({}, "k.d", kms_connect_callback=callback)
-        self.assertIs(opts._kms_connect_callback, callback)
-
-        for bad in [1, "not-callable", object()]:
-            with self.assertRaisesRegex(TypeError, "kms_connect_callback must be callable"):
-                AutoEncryptionOpts({}, "k.d", kms_connect_callback=bad)  # type: ignore[arg-type]
-
-        context = KMSConnectContext(host="kms.example.com", port=443, timeout=9.5)
-        self.assertEqual(context.host, "kms.example.com")
-        self.assertEqual(context.port, 443)
-        self.assertEqual(context.timeout, 9.5)
-        with self.assertRaises(dataclasses.FrozenInstanceError):
-            context.host = "evil.example.com"  # type: ignore[misc]
 
 
 # KMS connect callback unit and prose tests live in test_kms_connect.py.
